+++ android/app/src/main/java/com/yagubogu/ui/setting/component/dialog/LogoutDialog.kt
@@ -5,8 +5,8 @@ import androidx.compose.ui.Modifier
 import androidx.compose.ui.res.stringResource
 import androidx.compose.ui.tooling.preview.Preview
 import com.yagubogu.R
-import com.yagubogu.presentation.dialog.DefaultDialogUiModel
 import com.yagubogu.ui.common.component.DefaultDialog
+import com.yagubogu.ui.common.model.DefaultDialogUiModel
 
 @Composable
 fun LogoutDialog(
